Scripts/requirements.txt
google_api_python_client==2.123.0
google_auth_oauthlib==1.2.0
pandas==2.0.3
protobuf==5.26.1
